moves upsertCustomBg function to separate file
