+++ www/main/settings/utils/index.js
@@ -1,7 +1,4 @@
 export { settingsNavObj, settingsObj } from './parse-settings';
-export {
-  uploadImage,
-  upsertCustomBackgrounds,
-  removeCustomBackgroundFile,
-} from './theme-bg-uploader';
+export { uploadImage, removeCustomBackgroundFile } from './theme-bg-uploader';
 export { applyTheme } from './apply-theme';
+export { upsertCustomBackgrounds } from './theme-bg-reader';
